backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
pandas==0.25.3
numpy==1.19.5
nltk==3.5
spacy==3.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
textblob==0.15.3
readability==0.3.1
multiprocess==0.70.11.1
scipy==1.6.2
scikit-learn==0.24.2
mysql-connector-python==8.0.23
xlrd==1.2.0
click==7.1.1